+++ chevah/compat/tests/normal/test_filesystem.py
@@ -903,6 +903,108 @@ class TestLocalFilesystem(CompatTestCase, FilesystemTestMixin):
         self.assertTrue(isinstance(content[0], str))
         self.assertItemsEqual([folder_name, file_name], content)
 
+    def test_iterateFolderContent_not_found(self):
+        """
+        Raise OSError when trying to get folder for a non existent path.
+        """
+        segments = ['c', mk.string(), mk.string()]
+
+        with self.assertRaises(OSError) as context:
+            self.filesystem.iterateFolderContent(segments)
+
+        self.assertEqual(errno.ENOENT, context.exception.errno)
+
+    def test_iterateFolderContent_file(self):
+        """
+        Raise OSError when trying to get folder content for a file.
+        """
+        segments = self.fileInTemp()
+
+        with self.assertRaises(OSError) as context:
+            self.filesystem.iterateFolderContent(segments)
+
+        if self.os_family == 'nt':
+            # On Windows, we get a different error.
+            expected_error = errno.EINVAL
+        else:
+            expected_error = errno.ENOTDIR
+
+        self.assertEqual(expected_error, context.exception.errno)
+
+    def test_iterateFolderContent_empty(self):
+        """
+        Return empty iterator for empty folders.
+        """
+        segments = self.folderInTemp()
+
+        result = self.filesystem.iterateFolderContent(segments)
+
+        self.assertIteratorEqual([], result)
+
+    def test_iterateFolderContent_non_empty(self):
+        """
+        Return folder content as list of Unicode names.
+        """
+        base_segments = self.folderInTemp()
+        file_name = mk.makeFilename()
+        folder_name = mk.makeFilename()
+        file_segments = base_segments + [file_name]
+        folder_segments = base_segments + [folder_name]
+        mk.fs.createFile(file_segments)
+        mk.fs.createFolder(folder_segments)
+
+        content = self.filesystem.iterateFolderContent(base_segments)
+
+        result = list(content)
+        self.assertIsNotEmpty(result)
+        self.assertIsInstance(str, result[0])
+        self.assertItemsEqual([folder_name, file_name], result)
+
+    def test_iterateFolderContent_big(self):
+        """
+        It will not block on listing folders with many members.
+        """
+        # FIXME:4036:
+        # Enable full test once we have fast filesystem access.
+        if self.os_name == 'aix':
+            count = 3000
+            base_timeout = 0.02
+        elif self.os_name == 'hpux':
+            # HP-UX does not allow more than 32765 members in a folder
+            # and the slave is generally slow.
+            count = 32000
+            base_timeout = 0.1
+        else:
+            count = 35000
+            base_timeout = 0.1
+
+        base_segments = self.folderInTemp()
+
+        for i in range(count):
+            mk.fs.createFolder(base_segments + ['some-member-%s' % (i,)])
+
+        # We check that doing a direct listing will take a long time.
+        with self.assertRaises(AssertionError):
+            with self.assertExecutionTime(base_timeout):
+                result = self.filesystem.getFolderContent(base_segments)
+        self.assertEqual(count, len(result))
+
+        # Getting the iterator will not take long.
+        with self.assertExecutionTime(base_timeout):
+            iterator = self.filesystem.iterateFolderContent(base_segments)
+
+        # Iterating at any step will not take long.
+        result = []
+        try:
+            while True:
+                with self.assertExecutionTime(base_timeout):
+                    result.append(next(iterator))
+        except StopIteration:
+            """
+            We are at the end. All good.
+            """
+        self.assertEqual(count, len(result))
+
     def test_openFile_folder(self):
         """
         Raise OSError when trying to open a folder as file.