+++ paver.sh
@@ -61,12 +61,13 @@ OS='not-detected-yet'
 ARCH='x86'
 
 # Initialize default values from paver.conf
-PYTHON_VERSION='python2.7'
+PYTHON_CONFIGURATION='NOT-YET-DEFINED'
+PYTHON_VERSION='not.defined.yet'
+PYTHON_PLATFORM='unknown-os-and-arch'
+PYTHON_NAME='python2.7'
 BINARY_DIST_URI='https://binary.chevah.com/production'
 PIP_INDEX='http://pypi.chevah.com'
 PAVER_VERSION='1.2.1'
-PIP_VERSION="1.4.1.c4"
-SETUPTOOLS_VERSION="1.4.1"
 
 # Load repo specific configuration.
 source paver.conf
@@ -98,7 +99,13 @@ clean_build() {
     fi
     # In some case pip hangs with a build folder in temp and
     # will not continue until it is manually removed.
-    rm -rf /tmp/pip*
+    # On the OSX build server tmp is in $TMPDIR
+    if [ ! -z "${TMPDIR-}" ]; then
+        # check if TMPDIR is set before trying to clean it.
+        rm -rf ${TMPDIR}/pip*
+    else
+        rm -rf /tmp/pip*
+    fi
 }
 
 
@@ -153,57 +160,104 @@ execute() {
 # Update global variables with current paths.
 #
 update_path_variables() {
+    resolve_python_version
+
     if [ "${OS}" = "windows" ] ; then
         PYTHON_BIN="/lib/python.exe"
         PYTHON_LIB="/lib/Lib/"
     else
         PYTHON_BIN="/bin/python"
-        PYTHON_LIB="/lib/${PYTHON_VERSION}/"
+        PYTHON_LIB="/lib/${PYTHON_NAME}/"
     fi
 
     BUILD_FOLDER="build-${OS}-${ARCH}"
     PYTHON_BIN="${BUILD_FOLDER}${PYTHON_BIN}"
     PYTHON_LIB="${BUILD_FOLDER}${PYTHON_LIB}"
 
-    LOCAL_PYTHON_BINARY_DIST="$PYTHON_VERSION-$OS-$ARCH"
+    LOCAL_PYTHON_BINARY_DIST="$PYTHON_NAME-$OS-$ARCH"
 
     export PYTHONPATH=${BUILD_FOLDER}
 }
 
+#
+# Called to update the Python version env var based on the platform
+# advertised by the current environment.
+#
+resolve_python_version() {
+    local version_configuration=$PYTHON_CONFIGURATION
+    local candidate
+    local candidate_platform
+    local candidate_version
+
+    PYTHON_PLATFORM="$OS-$ARCH"
+
+    # This is a stupid way to iterate, up to 16 times while being OS neutral.
+    # We only support a maximum of 16 different versions.
+    for i in 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16; do
+        candidate=`echo ${version_configuration} | cut -d: -f$i`
+        if [ "$candidate" = "" ]; then
+            break
+        fi
+        candidate_platform=`echo $candidate | cut -d@ -f1`
+        candidate_version=`echo $candidate | cut -d@ -f2`
+
+        if [ "$candidate_platform" == "default" ]; then
+            # Set the default version in case no specific platform is
+            # configured.
+            PYTHON_VERSION=$candidate_version
+        fi
+
+        case $PYTHON_PLATFORM in
+            $candidate_platform*)
+                # We have a match for a specific platform, so we return
+                # as we don't want to look further.
+                PYTHON_VERSION=$candidate_version
+                return 0
+                ;;
+        esac
+
+    done
+}
 
 write_default_values() {
-    echo ${BUILD_FOLDER} ${PYTHON_VERSION} ${OS} ${ARCH} > DEFAULT_VALUES
+    echo ${BUILD_FOLDER} ${PYTHON_NAME} ${OS} ${ARCH} > DEFAULT_VALUES
 }
 
 
 #
-# Install brink package.
+# Install base package.
 #
-install_brink() {
+install_base_deps() {
+    local base_packages
+    base_packages="paver==$PAVER_VERSION"
     if [ "$BRINK_VERSION" = "skip" ]; then
         echo "Skipping brink installation."
-        return
+    else
+        base_packages="$base_packages chevah-brink==$BRINK_VERSION"
     fi
 
-    echo "Installing version: chevah-brink==$BRINK_VERSION of brink..."
+    echo "Installing $base_packages."
 
-    pip install "chevah-brink==$BRINK_VERSION"
+    pip_install "$base_packages"
 }
 
 
 #
-# Wrapper for python pip command.
-# * $1 - command name
-# * $2 - package_name and optional version.
+# Wrapper for python `pip install` command.
+# * $1 - package_name and optional version.
 #
-pip() {
+pip_install() {
     set +e
+    # There is a bug in pip/setuptools when using custom build folders.
+    # See https://github.com/pypa/pip/issues/3564
+    rm -rf ${BUILD_FOLDER}/pip-build
     ${PYTHON_BIN} -m \
-        pip.__init__ $1 $2 \
+        pip.__init__ install $1 \
+            --trusted-host pypi.chevah.com \
             --index-url=$PIP_INDEX/simple \
-            --download-cache=${CACHE_FOLDER} \
-            --find-links=file://${CACHE_FOLDER} \
-            --upgrade
+            --build=${BUILD_FOLDER}/pip-build \
+            --cache-dir=${CACHE_FOLDER} \
+            --use-wheel
 
     exit_code=$?
     set -e
@@ -219,9 +273,9 @@ pip() {
 #
 get_binary_dist() {
     local dist_name=$1
-    local remote_url=$2
+    local remote_base_url=$2
 
-    echo "Getting $dist_name from $remote_url..."
+    echo "Getting $dist_name from $remote_base_url..."
 
     tar_gz_file=${dist_name}.tar.gz
     tar_file=${dist_name}.tar
@@ -234,7 +288,8 @@ get_binary_dist() {
         rm -f $tar_gz_file
         rm -f $tar_file
         # Use 1M dot to reduce console pollution.
-        execute wget --progress=dot -e dotbytes=1M $remote_url/${tar_gz_file}
+        execute wget --progress=dot -e dotbytes=1M \
+            $remote_base_url/${tar_gz_file}
         execute gunzip $tar_gz_file
         execute tar -xf $tar_file
         rm -f $tar_gz_file
@@ -243,73 +298,111 @@ get_binary_dist() {
     popd
 }
 
+#
+# Check if we have a versioned Python distribution.
+#
+test_version_exists() {
+    local remote_base_url=$1
+    local wget_test
+    local target_file=python-${PYTHON_VERSION}-${OS}-${ARCH}.tar.gz
 
+    wget --spider $remote_base_url/${OS}/${ARCH}/$target_file
+    wget_test=$?
+    return $wget_test
+}
+
+#
+# Download and extract in cache the python distributable.
+#
+get_python_dist() {
+    local remote_base_url=$1
+    local python_distributable=python-${PYTHON_VERSION}-${OS}-${ARCH}
+    local wget_test
+
+    set +o errexit
+    test_version_exists $remote_base_url
+    wget_test=$?
+    set -o errexit
+
+    if [ $wget_test -eq 0 ]; then
+        # We have the requested python version.
+        get_binary_dist $python_distributable $remote_base_url/${OS}/${ARCH}
+    else
+        # Fall back to the non-versioned distribution.
+        echo "!!!Getting FALLBACK version!!!"
+        get_binary_dist $PYTHON_NAME-$OS-$ARCH $remote_base_url
+    fi
+}
+
+
+# copy_python can be called in a recursive way, and this is here to prevent
+# accidental infinite loops.
+COPY_PYTHON_RECURSIONS=0
 #
 # Copy python to build folder from binary distribution.
 #
 copy_python() {
 
-    local python_distributable="${CACHE_FOLDER}/${PYTHON_VERSION}-${OS}-${ARCH}"
-    local pip_package="pip-$PIP_VERSION"
-    local setuptools_package="setuptools-$SETUPTOOLS_VERSION"
+    local python_distributable="${CACHE_FOLDER}/${LOCAL_PYTHON_BINARY_DIST}"
+    local python_installed_version
+
+    COPY_PYTHON_RECURSIONS=`expr $COPY_PYTHON_RECURSIONS + 1`
+
+    if [ $COPY_PYTHON_RECURSIONS -gt 2 ]; then
+        echo "Too many calls to copy_python. ($COPY_PYTHON_RECURSIONS)"
+        exit 1
+    fi
 
     # Check that python dist was installed
     if [ ! -s ${PYTHON_BIN} ]; then
         # Install python-dist since everything else depends on it.
-        echo "Bootstrapping ${PYTHON_VERSION} environment to ${BUILD_FOLDER}..."
+        echo "Bootstrapping ${LOCAL_PYTHON_BINARY_DIST} environment" \
+            "to ${BUILD_FOLDER}..."
         mkdir -p ${BUILD_FOLDER}
 
         # If we don't have a cached python distributable,
         # get one together with default build system.
         if [ ! -d ${python_distributable} ]; then
-            echo "No ${PYTHON_VERSION} environment. Start downloading it..."
-            get_binary_dist \
-                ${PYTHON_VERSION}-${OS}-${ARCH} "$BINARY_DIST_URI/python"
+            echo "No ${LOCAL_PYTHON_BINARY_DIST} environment." \
+                "Start downloading it..."
+            get_python_dist "$BINARY_DIST_URI/python"
         fi
-        echo "Copying bootstrapping files... "
+        echo "Copying Python distribution files... "
         cp -R ${python_distributable}/* ${BUILD_FOLDER}
 
-        # Backwards compatibility with python 2.5 build.
-        if [[ "$PYTHON_VERSION" = "python2.5" ]]; then
-            # Copy include files.
-            if [ -d ${BUILD_FOLDER}/lib/config/include ]; then
-                cp -r ${BUILD_FOLDER}/lib/config/include ${BUILD_FOLDER}
-            fi
-
-            # Copy pywintypes25.dll as it is required by paver on windows.
-            if [ "$OS" = "windows" ]; then
-                cp -R ${BUILD_FOLDER}/lib/pywintypes25.dll . || true
-            fi
-        fi
-
-        if [ ! -d ${CACHE_FOLDER}/$pip_package ]; then
-            echo "No ${pip_package}. Start downloading it..."
-            get_binary_dist "$pip_package" "$PIP_INDEX/packages"
-        fi
-        rm -rf ${PYTHON_LIB}/site-packages/pip
-        cp -RL "${CACHE_FOLDER}/$pip_package/pip" ${PYTHON_LIB}/site-packages/
-
-        if [ ! -d ${CACHE_FOLDER}/$setuptools_package ]; then
-            echo "No ${setuptools_package}. Start downloading it..."
-            get_binary_dist "$setuptools_package" "$PIP_INDEX/packages"
-        fi
-        cp -RL "${CACHE_FOLDER}/$setuptools_package/setuptools" \
-            ${PYTHON_LIB}/site-packages/
-        cp -RL "${CACHE_FOLDER}/$setuptools_package/_markerlib" \
-            ${PYTHON_LIB}/site-packages/
-        cp -RL "${CACHE_FOLDER}/$setuptools_package/setuptools.egg-info" \
-            ${PYTHON_LIB}/site-packages/
-        cp "${CACHE_FOLDER}/$setuptools_package/pkg_resources.py" \
-            ${PYTHON_LIB}/site-packages/
-        cp -RL "${CACHE_FOLDER}/$setuptools_package/_markerlib" \
-            ${PYTHON_LIB}/site-packages/
-        cp "${CACHE_FOLDER}/$setuptools_package/easy_install.py" \
-            ${PYTHON_LIB}/site-package
-
-        # Once we have pip, we can use it.
-        pip install "paver==$PAVER_VERSION"
-
+        install_base_deps
         WAS_PYTHON_JUST_INSTALLED=1
+    else
+        # We have a Python, but we are not sure if is the right version.
+        local version_file=${BUILD_FOLDER}/lib/PYTHON_PACKAGE_VERSION
+        if [ -f $version_file ]; then
+            python_installed_version=`cat $version_file`
+            if [ "$PYTHON_VERSION" != "$python_installed_version" ]; then
+                # We have a different python installed.
+                # Remove it and try to install it again.
+                echo "Updating Python from" \
+                    $python_installed_version to $PYTHON_VERSION
+                rm -rf ${BUILD_FOLDER}
+                rm -rf ${python_distributable}
+                copy_python
+            fi
+        else
+            # The installed python has no version.
+            set +o errexit
+            test_version_exists "$BINARY_DIST_URI/python"
+            local test_version=$?
+            set -o errexit
+            if [ $test_version -eq 0 ]; then
+                echo "Updating Python from UNVERSIONED to $PYTHON_VERSION"
+                # We have a different python installed.
+                # Remove it and try to install it again.
+                rm -rf ${BUILD_FOLDER}
+                rm -rf ${python_distributable}
+                copy_python
+            else
+                echo "Leaving UNVERSIONED Python."
+            fi
+        fi
     fi
 
 }
@@ -323,8 +416,6 @@ install_dependencies(){
     if [ $WAS_PYTHON_JUST_INSTALLED -ne 1 ]; then
         return
     fi
-
-    install_brink
 
     set +e
     ${PYTHON_BIN} -c 'from paver.tasks import main; main()' deps
@@ -424,6 +515,16 @@ detect_os() {
 
         OS="solaris${os_version_chevah}"
 
+        # Solaris 10u8 (from 10/09) updated the libc version, so for older
+        # releases we build on 10u3, and use that up to 10u7 (from 5/09).
+        if [ "${OS}" = "solaris10" ]; then
+            # We extract the update number from the first line.
+            update=$(head -1 /etc/release | cut -d'_' -f2 | sed 's/[^0-9]*//g')
+            if [ "$update" -lt 8 ]; then
+                OS="solaris10u3"
+            fi
+        fi
+
     elif [ "${OS}" = "aix" ]; then
 
         ARCH="ppc$(getconf HARDWARE_BITMODE)"
@@ -497,8 +598,15 @@ detect_os() {
         os_version_raw=$(sw_vers -productVersion)
         check_os_version "Mac OS X" 10.8 "$os_version_raw" os_version_chevah
 
-        # For now, no matter the actual OS X version returned, we use '108'.
-        OS="osx108"
+        if [ ${os_version_chevah:0:2} -eq 10 -a \
+            ${os_version_chevah:2:2} -ge 12  ]; then
+            # For newer, macOS versions, we use '1012'.
+            OS="macos1012"
+        else
+            # For older, OS X versions, we use '108'.
+            OS="osx108"
+        fi
+
 
     elif [ "${OS}" = "freebsd" ]; then
         ARCH=$(uname -m)
@@ -514,9 +622,7 @@ detect_os() {
 
         os_version_raw=$(uname -r)
         check_os_version "OpenBSD" 5.9 "$os_version_raw" os_version_chevah
-
-        # For now, no matter the actual OpenBSD version returned, we use '59'.
-        OS="openbsd59"
+        OS="openbsd${os_version_chevah}"
 
     else
         echo 'Unsupported operating system:' $OS
@@ -558,7 +664,10 @@ if [ "$COMMAND" = "detect_os" ] ; then
 fi
 
 if [ "$COMMAND" = "get_python" ] ; then
-    get_binary_dist $2 "$BINARY_DIST_URI/python"
+    OS=$2
+    ARCH=$3
+    resolve_python_version
+    get_python_dist "$BINARY_DIST_URI/python"
     exit 0
 fi
 
@@ -572,17 +681,10 @@ write_default_values
 copy_python
 install_dependencies
 
-# Update while we migrate to wheels.
-# Should be removed later after we roll all new
-# python packages.
-setuptools_package="setuptools-$SETUPTOOLS_VERSION"
-cp -RL "${CACHE_FOLDER}/$setuptools_package/_markerlib" \
-    ${PYTHON_LIB}/site-packages/
-
 # Always update brink when running buildbot tasks.
 for paver_task in "deps" "test_os_dependent" "test_os_independent"; do
     if [ "$COMMAND" == "$paver_task" ] ; then
-        install_brink
+        install_base_deps
     fi
 done